# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: 1OF1 Trader Pro Affiliate Back Office API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /affiliate/apply:
    post:
      operationId: submitAffiliateApplication
      tags: [public]
      summary: Submit affiliate application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AffiliateApplicationRequest"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /r/{referralCode}:
    get:
      operationId: handleReferralRedirect
      tags: [public]
      summary: Handle referral redirect
      parameters:
        - name: referralCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "302":
          description: Redirect to landing page

  # Tracking
  /track/click:
    post:
      operationId: trackClick
      tags: [tracking]
      summary: Track referral click
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackClickRequest"
      responses:
        "200":
          description: Tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackClickResponse"
  /track/lead:
    post:
      operationId: trackLead
      tags: [tracking]
      summary: Track lead signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TrackLeadRequest"
      responses:
        "200":
          description: Tracked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # Webhooks
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    AffiliateApplicationRequest:
      type: object
      properties:
        fullName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        country:
          type: string
        telegram:
          type: string
          nullable: true
        discord:
          type: string
          nullable: true
        websiteUrl:
          type: string
          nullable: true
        twitterUrl:
          type: string
          nullable: true
        youtubeUrl:
          type: string
          nullable: true
        audienceType:
          type: string
        communitySize:
          type: string
        trafficSources:
          type: string
        whyJoin:
          type: string
        tradingExperience:
          type: string
        payoutMethod:
          type: string
          nullable: true
        payoutDetails:
          type: string
          nullable: true
      required:
        - fullName
        - email
        - phone
        - country
        - audienceType
        - communitySize
        - trafficSources
        - whyJoin
        - tradingExperience
...
    TrackClickRequest:
      type: object
      properties:
        affiliateCode:
          type: string
        ipAddress:
          type: string
          nullable: true
        userAgent:
          type: string
          nullable: true
        referrerUrl:
          type: string
          nullable: true
        landingPage:
          type: string
          nullable: true
        utmSource:
          type: string
          nullable: true
        utmMedium:
          type: string
          nullable: true
        utmCampaign:
          type: string
          nullable: true
        sessionId:
          type: string
          nullable: true
      required:
        - affiliateCode
    TrackClickResponse:
      type: object
      properties:
        clickId:
          type: integer
        sessionId:
          type: string
        affiliateCode:
          type: string
      required:
        - clickId
        - sessionId
        - affiliateCode
    TrackLeadRequest:
      type: object
      properties:
        sessionId:
          type: string
        email:
          type: string
          format: email
        fullName:
          type: string
          nullable: true
      required:
        - sessionId
        - email
...